+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/tests/model/IdentifierTypeName.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.health.qa.autotest.plr.util;
+package ca.bc.gov.health.qa.autotest.plr.web.tests.model;
 
 public enum IdentifierTypeName {
 	SELECT_ONE("Select One"),
