+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/tests/model/TelecommunicationType.java
@@ -1,13 +1,15 @@
-package ca.bc.gov.health.qa.autotest.plr.util;
+package ca.bc.gov.health.qa.autotest.plr.web.tests.model;
 
-public enum ElectronicAddressType {
-	EMAIL("E - Email"), 
-	FTP("F - FTP"),
-	HTTP("H - HTTP");
+public enum TelecommunicationType {
+	PHONE("T - Telephone"), 
+	MOBILE("MB - Mobile"),
+	PAGER("PG - Pager"),
+	FAX("FAX - Fax"),
+	MODEM("M - Modem");
 
 	private final String text;
 
-	ElectronicAddressType(String text) {
+	TelecommunicationType(String text) {
 		this.text = text;
 	}
 
@@ -21,8 +23,8 @@ public enum ElectronicAddressType {
 
 	public String getDataField() { return this.getEndText() + " (" + this.getStartText() + ")"; }
 
-	public static ElectronicAddressType fromString(String text) {
-		for (ElectronicAddressType b : ElectronicAddressType.values()) {
+	public static TelecommunicationType fromString(String text) {
+		for (TelecommunicationType b : TelecommunicationType.values()) {
 			if (b.text.equalsIgnoreCase(text)) {
 				return b;
 			}
